Add test to verify that all views are displayed
Verifies that all views are displayed in LearningActivity. This can be
useful when the tests run on the CI and that there is a failure that
comes from the fact that it couldn't find some view. At least then
you know that all views are displayed if the test passed.