Integrate the dummy dataset with the learning UI (#45)
* Integrate the dummy dataset with the learning UI

Also some refactoring because my separation between view and presenter
didn't make much sense.

* Refactor LearningActivity

Simplify targetOnDragListener by factorising two cases into the default
and factorise dragEntered/ExitedCallback because it was duplicate code,
as pointed out by CodeClimate.

* Add test to verify that all views are displayed

Verifies that all views are displayed in LearningActivity. This can be
useful when the tests run on the CI and that there is a failure that
comes from the fact that it couldn't find some view. At least then
you know that all views are displayed if the test passed.

Also refactored waitFor into TestUtils, as it may be useful for other
tests in the future.

* Adapt Learning UI to new DatasetInterface

* Make DummyCategory a data class
And adapt everything that broke as a consequence of this change

Co-authored-by: Niels Lachat <niels.lachat@epfl.ch>